Revert "Revert "Revert "Revert "Generic api code added.""""
This reverts commit 256dd5bf23f0fc9aa75620a6e957a20666861af0.

+++ Youtube MVVM Products/Screens/Product/Model/Product.swift
@@ -10,7 +10,7 @@ import Foundation
 struct Product: Decodable {
     let id: Int
     let title: String
-    let price: Float
+    let price: Double
     let description: String
     let category: String
     let image: String
@@ -18,6 +18,6 @@ struct Product: Decodable {
 }
 
 struct Rate: Decodable {
-    let rate: Float
+    let rate: Double
     let count: Int
 }
